+++ src/Regularization/Regularization.jl
@@ -5,3 +5,5 @@ include("KeplerValidation.jl")
 include("LeviCivita.jl")
 
 include("LeviCivitaOscillator.jl")
+
+include("SundmanTime.jl")